Move database to seperate module
